requirements.txt
cloud-volume
joblib
tqdm
pillow